# openapi/openapi.json
{
  "components": {
    "schemas": {
      "Command": {
        "enum": [
          "start",
          "pause",
          "resume",
          "skip_phase",
          "extend_phase",
          "complete_phase",
          "cancel",
          "restart"
        ],
        "title": "Command",
        "type": "string"
      },
      "CommandInput": {
        "properties": {
          "command": {
            "$ref": "#/components/schemas/Command"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "extend_seconds": {
            "anyOf": [
              {
                "maximum": 14400.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extend Seconds"
          }
        },
        "required": [
          "command",
          "expected_version"
        ],
        "title": "CommandInput",
        "type": "object"
      },
      "FamilyCreate": {
        "properties": {
          "description": {
            "default": "",
            "maxLength": 2000,
            "title": "Description",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "FamilyCreate",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PhaseInput": {
        "properties": {
          "duration_seconds": {
            "maximum": 14400.0,
            "minimum": 1.0,
            "title": "Duration Seconds",
            "type": "integer"
          },
          "extendable": {
            "default": true,
            "title": "Extendable",
            "type": "boolean"
          },
          "key": {
            "maxLength": 60,
            "minLength": 1,
            "pattern": "^[a-z0-9_-]+$",
            "title": "Key",
            "type": "string"
          },
          "max_duration_seconds": {
            "default": 14400,
            "maximum": 86400.0,
            "minimum": 1.0,
            "title": "Max Duration Seconds",
            "type": "integer"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "phase_type": {
            "$ref": "#/components/schemas/PhaseType"
          },
          "skippable": {
            "default": true,
            "title": "Skippable",
            "type": "boolean"
          }
        },
        "required": [
          "key",
          "name",
          "phase_type",
          "duration_seconds"
        ],
        "title": "PhaseInput",
        "type": "object"
      },
      "PhaseType": {
        "enum": [
          "focus",
          "short_break",
          "long_break",
          "custom"
        ],
        "title": "PhaseType",
        "type": "string"
      },
      "SessionCreate": {
        "properties": {
          "family_version_id": {
            "format": "uuid",
            "title": "Family Version Id",
            "type": "string"
          }
        },
        "required": [
          "family_version_id"
        ],
        "title": "SessionCreate",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VersionCreate": {
        "properties": {
          "phases": {
            "items": {
              "$ref": "#/components/schemas/PhaseInput"
            },
            "maxItems": 32,
            "minItems": 1,
            "title": "Phases",
            "type": "array"
          }
        },
        "required": [
          "phases"
        ],
        "title": "VersionCreate",
        "type": "object"
      }
    },
    "securitySchemes": {
      "ModorocoApiKey": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Modoroco API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/families": {
      "get": {
        "operationId": "list_families_v1_families_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Families V1 Families Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "ModorocoApiKey": []
          }
        ],
        "summary": "List Families"
      },
      "post": {
        "operationId": "create_family_v1_families_post",
        "parameters": [
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Family V1 Families Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ModorocoApiKey": []
          }
        ],
        "summary": "Create Family"
      }
    },
    "/v1/families/{family_id}": {
      "get": {
        "operationId": "read_family_v1_families__family_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "family_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Family Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Read Family V1 Families  Family Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ModorocoApiKey": []
          }
        ],
        "summary": "Read Family"
      }
    },
    "/v1/families/{family_id}/versions": {
      "post": {
        "operationId": "publish_version_v1_families__family_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "family_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Family Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Publish Version V1 Families  Family Id  Versions Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ModorocoApiKey": []
          }
        ],
        "summary": "Publish Version"
      }
    },
    "/v1/sessions": {
      "post": {
        "operationId": "create_session_v1_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Session V1 Sessions Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ModorocoApiKey": []
          }
        ],
        "summary": "Create Session"
      }
    },
    "/v1/sessions/{session_id}": {
      "get": {
        "operationId": "read_session_v1_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Read Session V1 Sessions  Session Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ModorocoApiKey": []
          }
        ],
        "summary": "Read Session"
      }
    },
    "/v1/sessions/{session_id}/commands": {
      "post": {
        "operationId": "command_session_v1_sessions__session_id__commands_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Command Session V1 Sessions  Session Id  Commands Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ModorocoApiKey": []
          }
        ],
        "summary": "Command Session"
      }
    },
    "/v1/sessions/{session_id}/events": {
      "get": {
        "operationId": "session_events_v1_sessions__session_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Session Events V1 Sessions  Session Id  Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ModorocoApiKey": []
          }
        ],
        "summary": "Session Events"
      }
    },
    "/v1/sessions/{session_id}/stream": {
      "get": {
        "operationId": "session_stream_v1_sessions__session_id__stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ModorocoApiKey": []
          }
        ],
        "summary": "Session Stream"
      }
    }
  }
}